Improvements regarding code review
