+++ Kitodo/src/main/java/org/kitodo/production/forms/dataeditor/MediaPartialsPanel.java
@@ -11,22 +11,21 @@
 
 package org.kitodo.production.forms.dataeditor;
 
+import static org.kitodo.production.helper.metadata.MediaPartialHelper.calculateExtentAndSortMediaPartials;
+import static org.kitodo.production.helper.metadata.MediaPartialHelper.convertFormattedTimeToMilliseconds;
+
+import javax.faces.context.FacesContext;
+import javax.faces.model.SelectItem;
 import java.io.Serializable;
 import java.net.URI;
 import java.util.ArrayList;
 import java.util.Collection;
-import java.util.Comparator;
 import java.util.LinkedHashMap;
 import java.util.List;
-import java.util.ListIterator;
 import java.util.Map;
 import java.util.Objects;
-import java.util.concurrent.TimeUnit;
 import java.util.regex.Pattern;
 import java.util.stream.Collectors;
-
-import javax.faces.context.FacesContext;
-import javax.faces.model.SelectItem;
 
 import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.lang3.tuple.Pair;
@@ -37,13 +36,13 @@ import org.kitodo.api.dataformat.MediaVariant;
 import org.kitodo.api.dataformat.PhysicalDivision;
 import org.kitodo.api.dataformat.View;
 import org.kitodo.production.helper.Helper;
+import org.kitodo.production.helper.metadata.MediaPartialHelper;
 import org.kitodo.production.services.dataeditor.DataEditorService;
 import org.kitodo.utils.MediaUtil;
 import org.omnifaces.util.Ajax;
 
 public class MediaPartialsPanel implements Serializable {
 
-    public static final String FORMATTED_TIME_REGEX = "([0-1]\\d|2[0-3]):[0-5]\\d:[0-5]\\d";
     public static final String REQUEST_PARAMETER_MEDIA_DURATION = "mediaDuration";
     public static final String[] UPDATE_CLIENT_IDENTIFIERS = {"structureTreeForm",
                                                               "imagePreviewForm:mediaDetailMediaPartialsContainer",
@@ -96,7 +95,7 @@ public class MediaPartialsPanel implements Serializable {
         String error = null;
         if (StringUtils.isEmpty(getMediaDuration())) {
             error = "mediaPartialFormMediaDurationEmpty";
-        } else if (!Pattern.compile(MediaPartialsPanel.FORMATTED_TIME_REGEX).matcher(getMediaDuration()).matches()) {
+        } else if (!Pattern.compile(MediaPartialHelper.FORMATTED_TIME_REGEX).matcher(getMediaDuration()).matches()) {
             error = "mediaPartialFormMediaDurationWrongTimeFormat";
         }
         return error;
@@ -164,7 +163,8 @@ public class MediaPartialsPanel implements Serializable {
             Collection<String> mediaPartialDivisionIds = dataEditor.getRulesetManagement()
                     .getFunctionalDivisions(FunctionalDivision.MEDIA_PARTIAL);
             mediaPartialDivisions = mediaPartialDivisions.stream()
-                    .filter(selectItem -> mediaPartialDivisionIds.contains(selectItem.getValue()))
+                    .filter(selectItem -> selectItem.getValue() instanceof String)
+                    .filter(selectItem -> mediaPartialDivisionIds.contains((String) selectItem.getValue()))
                     .collect(Collectors.toList());
         }
         return mediaPartialDivisions;
@@ -197,95 +197,6 @@ public class MediaPartialsPanel implements Serializable {
         if (params.containsKey(REQUEST_PARAMETER_MEDIA_DURATION)) {
             mediaDuration = params.get(REQUEST_PARAMETER_MEDIA_DURATION);
         }
-    }
-
-    /**
-     * Calculates the extent field of every media partial and sort media partials by begin.
-     *
-     * @param logicalDivisions
-     *         The logical divisions of media partials.
-     * @param mediaDuration
-     *         The media duration.
-     */
-    public static void calculateExtentAndSortMediaPartials(List<LogicalDivision> logicalDivisions, Long mediaDuration) {
-        calculateExtentForMediaPartials(logicalDivisions, mediaDuration);
-        logicalDivisions.sort(getLogicalDivisionComparator());
-    }
-
-    /**
-     * Calculate the extent of a media partial.
-     *
-     * Calculates the duration or extent of a media partial until the next one or until the end of the media.
-     *
-     * @param logicalDivisions
-     *         The logical divisions of media partials.
-     * @param mediaDuration
-     *         The media duration.
-     */
-    private static void calculateExtentForMediaPartials(List<LogicalDivision> logicalDivisions, Long mediaDuration) {
-        // sorting reverse by begin
-        logicalDivisions.sort(getLogicalDivisionComparator().reversed());
-
-        ListIterator<LogicalDivision> iterator = logicalDivisions.listIterator();
-        LogicalDivision previousLogicalDivision = null;
-        while (iterator.hasNext()) {
-            LogicalDivision logicalDivision = iterator.next();
-            MediaPartialView mediaPartialView = logicalDivision.getViews().getFirst().getPhysicalDivision()
-                    .getMediaPartialView();
-            if (Objects.nonNull(previousLogicalDivision)) {
-                // calculate the duration of media partial to previous media partial
-                PhysicalDivision previousPhysicalDivision = previousLogicalDivision.getViews().getFirst()
-                        .getPhysicalDivision();
-                if (previousPhysicalDivision.hasMediaPartialView()) {
-                    mediaPartialView.setExtent(convertMillisecondsToFormattedTime(convertFormattedTimeToMilliseconds(
-                            previousPhysicalDivision.getMediaPartialView()
-                                    .getBegin()) - convertFormattedTimeToMilliseconds(
-                            mediaPartialView.getBegin())));
-                }
-            } else {
-                // calculate the duration of media partial to the end of media
-                mediaPartialView.setExtent(convertMillisecondsToFormattedTime(
-                        mediaDuration - convertFormattedTimeToMilliseconds(mediaPartialView.getBegin())));
-            }
-            previousLogicalDivision = logicalDivision;
-        }
-    }
-
-    /**
-     * Convert formatted time to milliseconds.
-     *
-     * @param formattedTime
-     *         The formatted time in form of {@value #FORMATTED_TIME_REGEX}
-     * @return The milliseconds
-     */
-    public static Long convertFormattedTimeToMilliseconds(String formattedTime) {
-        if (formattedTime.contains(".")) {
-            formattedTime = formattedTime.split(".")[0];
-        }
-        String[] time = formattedTime.split(":");
-        return Long.valueOf(
-                Integer.valueOf(time[0]) * 3600 + Integer.valueOf(time[1]) * 60 + Integer.valueOf(time[2])) * 1000;
-    }
-
-    public static String convertMillisecondsToFormattedTime(Long milliseconds) {
-        return String.format("%02d:%02d:%02d", TimeUnit.MILLISECONDS.toHours(milliseconds),
-                TimeUnit.MILLISECONDS.toMinutes(milliseconds) % 60, TimeUnit.MILLISECONDS.toSeconds(milliseconds) % 60);
-    }
-
-    private static Comparator<LogicalDivision> getLogicalDivisionComparator() {
-        return (logicalDivisionA, logicalDivisionB) -> {
-            View viewA = logicalDivisionA.getViews().getFirst();
-            View viewB = logicalDivisionB.getViews().getFirst();
-            if (Objects.nonNull(viewA) && Objects.nonNull(viewB)) {
-                PhysicalDivision physicalDivisionA = viewA.getPhysicalDivision();
-                PhysicalDivision physicalDivisionB = viewB.getPhysicalDivision();
-                if (physicalDivisionA.hasMediaPartialView() && physicalDivisionB.hasMediaPartialView()) {
-                    return physicalDivisionA.getMediaPartialView().getBegin()
-                            .compareTo(physicalDivisionB.getMediaPartialView().getBegin());
-                }
-            }
-            return Integer.compare(logicalDivisionA.getOrder(), logicalDivisionB.getOrder());
-        };
     }
 
     /**
